added 1 output parameter to sceNpInit test data

+++ les_psn.cpp
@@ -11,6 +11,6 @@ void LES_sceNpInit(int a)
 	int r;
 	LES_FUNCTION_START(sceNpInit, void)
 	LES_FUNCTION_INPUTS_1(int, a)
-	LES_FUNCTION_OUTPUTS_1(int, r)
+	LES_FUNCTION_OUTPUTS_1(short, r)
 	fprintf(stderr, "END\n");
 }
